feat: implementing post, patch and get in Person controller

+++ back-end/src/main/java/com/pms/calprop/dto/PersonRequest.java
@@ -2,6 +2,6 @@ package com.pms.calprop.dto;
 
 import java.math.BigDecimal;
 
-public record PersonRequest(String name, BigDecimal salary, BigDecimal reservePercentage) {
+public record PersonRequest(String name, BigDecimal salary, Double reservePercentage) {
 
 }
